fix code warning

+++ frameworks/native/tls_socket/src/tls_socket_server.cpp
@@ -1411,11 +1411,11 @@ void TLSSocketServer::PollThread(const TlsSocket::TLSConnectOptions &tlsListenOp
             for (int i = 0; i < g_userCounter + 1; ++i) {
                 if ((fds_[i].fd == listenSocketFd_) && (static_cast<uint16_t>(fds_[i].revents) & POLLIN)) {
                     ProcessTcpAccept(tlsOption, ++clientId);
-                } else if ((fds_[i].revents & POLLRDHUP) || (fds_[i].revents & POLLERR)) {
+		 } else if ((static_cast<uint16_t>(fds_[i].revents) & POLLRDHUP) || (static_cast<uint16_t>(fds_[i].revents) & POLLERR)) {
                     RemoveConnect(fds_[i].fd);
                     DropFdFromPollList(i);
                     NETSTACK_LOGI("A client left");
-                } else if (fds_[i].revents & POLLIN) {
+                } else if (static_cast<uint16_t>(fds_[i].revents) & POLLIN) {
                     RecvRemoteInfo(fds_[i].fd, i);
                 }
             }
